refactor(autoreview): Move ExtensionPoint to another location
