+++ tests/Architecture/PHPat/Selector/InfectionSelector.php
@@ -36,6 +36,7 @@ declare(strict_types=1);
 namespace Infection\Tests\Architecture\PHPat\Selector;
 
 use Infection\CannotBeInstantiated;
+use Infection\Tests\Architecture\PHPat\Selector\Support\ExtensionPoint;
 use PHPat\Selector\ClassImplements;
 use PHPat\Selector\Selector;
 use PHPat\Selector\SelectorInterface;
